requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
jinja2>=3.0.0
python-dotenv>=0.20.0
boto3>=1.26.0 # Added boto3